+++ packages/fscomponents/src/components/__stories__/Carousel.story.tsx
@@ -13,8 +13,8 @@ import {
 import { storiesOf } from '@storybook/react'; // tslint:disable-line:no-implicit-dependencies
 import {
   boolean
-// tslint:disable-next-line no-submodule-imports no-implicit-dependencies
-} from '@storybook/addon-knobs/react';
+// tslint:disable-next-line no-implicit-dependencies
+} from '@storybook/addon-knobs';
 
 import { Carousel } from '../Carousel';