+++ packages/fscomponents/src/components/__stories__/Form.story.tsx
@@ -6,11 +6,12 @@ import {
   color,
   object,
   text
-// tslint:disable-next-line no-submodule-imports no-implicit-dependencies
-} from '@storybook/addon-knobs/react';
+// tslint:disable-next-line no-implicit-dependencies
+} from '@storybook/addon-knobs';
 // @ts-ignore: TODO fix types for tcomb-form-native
 import * as t from 'tcomb-form-native';
-import { Form, FormLabelPosition } from '../Form/Form';
+import { Form } from '../Form/Form';
+import { FormLabelPosition } from '../Form/Templates';
 import { Button } from '../Button';
 import { emailRegex } from '../../lib/email';